+++ querys/stores procedures/Registros, ediciones y eliminaciones/producto/sp_registrar_producto.sql
@@ -16,7 +16,7 @@ AS
 BEGIN
     SET NOCOUNT ON;
 
-    DECLARE @nuevo_id INT;
+    DECLARE @nuevo_id BIGINT;
 
     BEGIN TRY
         BEGIN TRANSACTION;
